+++ src/main/java/com/factual/driver/Constants.java
@@ -85,4 +85,8 @@ public class Constants {
   protected static final String INSERT_FACTUAL_ID = "factual_id";
   protected static final String INSERT_NEW_ENTITY = "new_entity";
 
+  // Clear
+  protected static final String CLEAR_FIELDS = "fields";
+  protected static final String CLEAR_FACTUAL_ID = "factual_id";
+
 }